voxel_t = uint8_t: 1.46 s

+++ Common.hpp
@@ -94,7 +94,7 @@ typedef union
   int32_t int32[4];
 } union128;
 
-typedef uint32_t voxel_t;
+typedef uint8_t voxel_t;
 #define WORD (sizeof(voxel_t)*8)
 
 typedef int16_t mol_t;
